refactor: split playlist item mutation/undo into playlist/mutation_undo.rs (#577 item 18)
Completes item 18's playlist.rs split: item-level CRUD
(get_playlist_tracks, add/remove/dedup/trim/reorder*/clear_playlist)
and the undo/redo stack move into playlist/mutation_undo.rs, as a
second `impl PlaylistManager` block alongside the one in playlist.rs.

playlist.rs is now 389 lines, holding just top-level playlist CRUD
(new/create/rename/delete/get_playlists/queue/replace_queue) plus the
PlaylistManager/ClearedItem/PlaylistOp struct definitions —
ClearedItem/PlaylistOp stay here (not moved into mutation_undo.rs)
since PlaylistOp is the type of PlaylistManager's own undo_stack/
redo_stack fields, and a private type can't live in a child module
while being named in the parent's struct definition. renumber_positions
stays too, since it's shared by both mutation_undo.rs and dynamic.rs
(added in the prior PR in this series).

No behavior change — pure move. 218/219 tests pass throughout,
clippy/fmt clean (same 12 pre-existing unrelated findings in
player.rs/tags.rs as prior PRs in this series).
